requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.1
sqlparse==0.2.4
